requirements.txt
torch==2.11.0
torchvision==0.26.0
eva-decord==0.6.1
diffusers==0.33.1
Pillow==9.2
einops==0.8.0
scipy==1.14.0
numpy==1.26.4
opencv-python==4.10.0.84
omegaconf
einops
accelerate==0.30.1
moviepy==1.0.3
hydra-core
pytest
pydantic<2.10
gradio>=4.0.0
iopaint
